ProvisioningManagerTest fixed
